fix: needWeight 자료형 수정

+++ molaeng_back/src/main/java/idle/molaeng_back/recipe/model/entity/RecipeIngredient.java
@@ -18,7 +18,7 @@ public class RecipeIngredient {
     private long recipeIngredientId;
 
     @Column(name="need_weight")
-    private float needWeight;
+    private double needWeight;
 
     @Column(name="weight_unit")
     private String weightUnit;
